requirements.txt
locust
mysql-connector-python
